github page setup
not the version that was initially uploaded to the game jam because i forgot to clone it.

however, most of the differences are bug fixes + visual changes.  the only "big" change is that Request Cards go to random pre-determined spaces instead of anywhere in the play area.

all changes are included in _patch-notes.txt